+++ weblate/trans/views/basic.py
@@ -34,15 +34,17 @@ from weblate.trans.forms import (
     AnnouncementForm,
     AutoForm,
     BulkEditForm,
+    ComponentDeleteForm,
     ComponentMoveForm,
     ComponentRenameForm,
-    DeleteForm,
     DownloadForm,
     NewUnitForm,
+    ProjectDeleteForm,
     ProjectRenameForm,
     ReplaceForm,
     ReportsForm,
     SearchForm,
+    TranslationDeleteForm,
     get_new_language_form,
     get_upload_form,
 )
@@ -148,7 +150,7 @@ def show_project(request, project):
                 AnnouncementForm, user, "project.edit", obj
             ),
             "delete_form": optional_form(
-                DeleteForm, user, "project.edit", obj, obj=obj
+                ProjectDeleteForm, user, "project.edit", obj, obj=obj
             ),
             "rename_form": optional_form(
                 ProjectRenameForm,
@@ -215,7 +217,7 @@ def show_component(request, project, component):
                 AnnouncementForm, user, "component.edit", obj
             ),
             "delete_form": optional_form(
-                DeleteForm, user, "component.edit", obj, obj=obj
+                ComponentDeleteForm, user, "component.edit", obj, obj=obj
             ),
             "rename_form": optional_form(
                 ComponentRenameForm,
@@ -282,7 +284,7 @@ def show_translation(request, project, component, lang):
                 AnnouncementForm, user, "component.edit", obj
             ),
             "delete_form": optional_form(
-                DeleteForm, user, "translation.delete", obj, obj=obj
+                TranslationDeleteForm, user, "translation.delete", obj, obj=obj
             ),
             "last_changes": last_changes,
             "last_changes_url": urlencode(obj.get_reverse_url_kwargs()),